# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: AMACTIVE API
  description: >
    API de controle de estoque e vendas da AMACTIVE (moda fitness feminina).
    Uso interno — sem emissão fiscal e sem gateway de pagamento real no MVP.
    Ver docs/SDD.md para o desenho arquitetural completo.
  version: "0.1.0"
  contact:
    name: AMACTIVE Engineering
servers:
  - url: http://localhost:8000
    description: Ambiente local (docker-compose)
...
security:
  - bearerAuth: []
paths:
...
  /categorias:
    get:
      tags: [Categorias]
      summary: Listar categorias
      operationId: listarCategorias
      responses:
        "200":
          description: Lista de categorias
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoriaListResponse"
    post:
      tags: [Categorias]
      summary: Criar categoria
      operationId: criarCategoria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarCategoriaRequest"
      responses:
        "201":
          description: Categoria criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoriaResponse"
        "422":
          $ref: "#/components/responses/ErroValidacao"

  # ───────────────────────── Produtos ─────────────────────────
  /produtos:
    get:
      tags: [Produtos]
      summary: Listar produtos
      operationId: listarProdutos
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: busca
          in: query
          description: Busca textual por nome do produto
          schema: { type: string }
        - name: categoria_id
          in: query
          schema: { type: string, format: uuid }
        - name: ativo
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Lista paginada de produtos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProdutoListResponse"
    post:
      tags: [Produtos]
      summary: Criar produto
      operationId: criarProduto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarProdutoRequest"
      responses:
        "201":
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProdutoResponse"
        "422":
          $ref: "#/components/responses/ErroValidacao"
  /produtos/{produtoId}:
    parameters:
      - $ref: "#/components/parameters/ProdutoId"
    get:
      tags: [Produtos]
      summary: Obter produto por id (inclui variantes)
      operationId: obterProduto
      responses:
        "200":
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProdutoDetalheResponse"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    put:
      tags: [Produtos]
      summary: Atualizar produto
      operationId: atualizarProduto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarProdutoRequest"
      responses:
        "200":
          description: Produto atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProdutoResponse"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "422":
          $ref: "#/components/responses/ErroValidacao"
    delete:
      tags: [Produtos]
      summary: Inativar produto (soft delete — ativo=false)
      operationId: inativarProduto
      responses:
        "204":
          description: Produto inativado
        "404":
          $ref: "#/components/responses/NaoEncontrado"

  # ───────────────────────── Variantes ─────────────────────────
  /produtos/{produtoId}/variantes:
    parameters:
      - $ref: "#/components/parameters/ProdutoId"
    get:
      tags: [Variantes]
      summary: Listar variantes (SKUs) de um produto
      operationId: listarVariantesDoProduto
      responses:
        "200":
          description: Lista de variantes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/VarianteResponse" }
    post:
      tags: [Variantes]
      summary: Criar variante (SKU) para um produto
      operationId: criarVariante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarVarianteRequest"
      responses:
        "201":
          description: Variante criada (estoque inicial zerado é criado automaticamente)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VarianteResponse"
        "422":
          $ref: "#/components/responses/ErroValidacao"
  /variantes/{varianteId}:
    parameters:
      - $ref: "#/components/parameters/VarianteId"
    get:
      tags: [Variantes]
      summary: Obter variante por id
      operationId: obterVariante
      responses:
        "200":
          description: Variante encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VarianteResponse"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    put:
      tags: [Variantes]
      summary: Atualizar variante (preço, cor, tamanho)
      operationId: atualizarVariante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarVarianteRequest"
      responses:
        "200":
          description: Variante atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VarianteResponse"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    delete:
      tags: [Variantes]
      summary: Inativar variante (soft delete — ativo=false)
      operationId: inativarVariante
      responses:
        "204":
          description: Variante inativada
        "404":
          $ref: "#/components/responses/NaoEncontrado"

  # ───────────────────────── Imagens ─────────────────────────
  /produtos/{produtoId}/imagens:
    parameters:
      - $ref: "#/components/parameters/ProdutoId"
    get:
      tags: [Imagens]
      summary: Listar imagens (galeria) de um produto
      operationId: listarImagensDoProduto
      parameters:
        - name: cor
          in: query
          description: Filtra a galeria por cor (mesmo valor livre usado em produto_variante.cor)
          schema: { type: string }
      responses:
        "200":
          description: Galeria de imagens do produto, ordenada por `cor` e `ordem`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagemListResponse"
    post:
      tags: [Imagens]
      summary: Upload de imagem de produto
      description: >
        A `cor` informada precisa corresponder a uma cor existente entre as
        variantes ativas do produto (senão 422 — `cor` não é FK, ver
        docs/data-model.md decisão #13). Aceita apenas `image/jpeg`,
        `image/png` e `image/webp`, até 5MB. A primeira imagem enviada para
        uma cor é automaticamente marcada como `principal`.
      operationId: uploadImagem
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [cor, arquivo]
              properties:
                cor: { type: string, maxLength: 50 }
                arquivo: { type: string, format: binary }
      responses:
        "201":
          description: Imagem criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagemResponse"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "422":
          $ref: "#/components/responses/ErroValidacao"
        "413":
          description: Arquivo excede o tamanho máximo permitido (5MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /produtos/{produtoId}/imagens/{imagemId}:
    parameters:
      - $ref: "#/components/parameters/ProdutoId"
      - $ref: "#/components/parameters/ImagemId"
    patch:
      tags: [Imagens]
      summary: Reordenar imagem (atualizar `ordem`)
      operationId: atualizarOrdemImagem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarOrdemImagemRequest"
      responses:
        "200":
          description: Imagem atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagemResponse"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "422":
          $ref: "#/components/responses/ErroValidacao"
    delete:
      tags: [Imagens]
      summary: Remover imagem (registro em banco + arquivo físico em disco)
      operationId: removerImagem
      responses:
        "204":
          description: Imagem removida
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /produtos/{produtoId}/imagens/{imagemId}/principal:
    parameters:
      - $ref: "#/components/parameters/ProdutoId"
      - $ref: "#/components/parameters/ImagemId"
    patch:
      tags: [Imagens]
      summary: Definir imagem como principal (capa) da sua cor
      description: >
        Desmarca atomicamente a imagem principal anterior da mesma cor —
        no máximo uma principal por produto+cor (índice único parcial em
        `produto_imagem`, ver docs/data-model.md decisão #13).
      operationId: definirImagemPrincipal
      responses:
        "200":
          description: Imagem definida como principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagemResponse"
        "404":
          $ref: "#/components/responses/NaoEncontrado"

  # ───────────────────────── Estoque ─────────────────────────
  /estoque:
    get:
      tags: [Estoque]
      summary: Listar saldo de estoque por variante
      operationId: listarEstoque
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: sku
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lista paginada de saldos de estoque
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstoqueListResponse"
...
  /estoque/movimentacoes:
    get:
      tags: [Estoque]
      summary: Listar histórico de movimentações de estoque
      operationId: listarMovimentacoes
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: variante_id
          in: query
          schema: { type: string, format: uuid }
        - name: tipo
          in: query
          schema: { $ref: "#/components/schemas/TipoMovimentacao" }
      responses:
        "200":
          description: Lista paginada de movimentações
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovimentacaoListResponse"
    post:
      tags: [Estoque]
      summary: Registrar movimentação manual de estoque (entrada, saída ou ajuste)
      description: >
        Movimentações originadas de uma venda são criadas automaticamente pelo
        contexto de Vendas ao confirmar um Pedido (ver POST /pedidos) — este
        endpoint é para movimentações manuais (compra, ajuste de inventário,
        devolução, perda).
      operationId: criarMovimentacao
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarMovimentacaoRequest"
      responses:
        "201":
          description: Movimentação registrada e saldo de estoque atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovimentacaoResponse"
        "422":
          description: Dados inválidos ou saldo insuficiente para SAIDA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  # ───────────────────────── Pedidos (Vendas) ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    ProdutoId:
      name: produtoId
      in: path
      required: true
      schema: { type: string, format: uuid }
    VarianteId:
      name: varianteId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NaoEncontrado:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    ErroValidacao:
      description: Dados de entrada inválidos
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    CriarCategoriaRequest:
      type: object
      required: [nome]
      properties:
        nome: { type: string, maxLength: 100 }
    CategoriaResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }
        slug: { type: string }
        ativo: { type: boolean }
    CategoriaListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/CategoriaResponse" }

    # ── Produtos ──
    CriarProdutoRequest:
      type: object
      required: [nome]
      properties:
        nome: { type: string, maxLength: 200 }
        descricao: { type: string }
        categoria_id: { type: string, format: uuid, nullable: true }
        marca: { type: string, default: "AMACTIVE" }
    AtualizarProdutoRequest:
      allOf:
        - $ref: "#/components/schemas/CriarProdutoRequest"
        - type: object
          properties:
            ativo: { type: boolean }
    ProdutoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }
        descricao: { type: string, nullable: true }
        categoria_id: { type: string, format: uuid, nullable: true }
        marca: { type: string }
        ativo: { type: boolean }
        criado_em: { type: string, format: date-time }
    ProdutoDetalheResponse:
      allOf:
        - $ref: "#/components/schemas/ProdutoResponse"
        - type: object
          properties:
            variantes:
              type: array
              items: { $ref: "#/components/schemas/VarianteResponse" }
    ProdutoListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/ProdutoResponse" }
        pagination: { $ref: "#/components/schemas/Pagination" }

    # ── Variantes ──
...
    CriarVarianteRequest:
      type: object
      required: [tamanho, cor, preco_venda]
      properties:
        sku: { type: string, description: "Se omitido, é gerado automaticamente" }
        tamanho: { $ref: "#/components/schemas/Tamanho" }
        cor: { type: string, maxLength: 50 }
        preco_venda: { type: string, pattern: '^\d+\.\d{2}$', example: "129.90" }
        preco_custo: { type: string, pattern: '^\d+\.\d{2}$', nullable: true }
        estoque_inicial:
          type: integer
          minimum: 0
          default: 0
          description: Quantidade inicial de estoque, gera uma movimentação ENTRADA/AJUSTE_INVENTARIO automática
    AtualizarVarianteRequest:
      type: object
      properties:
        cor: { type: string }
        preco_venda: { type: string, pattern: '^\d+\.\d{2}$' }
        preco_custo: { type: string, pattern: '^\d+\.\d{2}$', nullable: true }
        ativo: { type: boolean }
    VarianteResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        produto_id: { type: string, format: uuid }
        sku: { type: string }
        tamanho: { $ref: "#/components/schemas/Tamanho" }
        cor: { type: string }
        preco_venda: { type: string }
        preco_custo: { type: string, nullable: true }
        ativo: { type: boolean }
        quantidade_estoque: { type: integer, description: "Saldo atual (denormalizado para conveniência)" }

    # ── Imagens ──
    ImagemResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        produto_id: { type: string, format: uuid }
        cor: { type: string }
        url: { type: string, description: "Caminho público, ex: /media/produtos/{produtoId}/{uuid}.jpg" }
        ordem: { type: integer, minimum: 0 }
        principal: { type: boolean }
        criado_em: { type: string, format: date-time }
    ImagemListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/ImagemResponse" }
...